Side Task 3 Fix, Swagger Improvements, Bugfixes

+++ pkg/repository/user_postgres.go
@@ -21,14 +21,32 @@ func (r *UserPostgres) GetUserSegments(usr_id int) ([]string, error) {
 	var id int
 
 	// update user segments
-	query := fmt.Sprintf("DELETE FROM %s WHERE user_id = $1 AND seg_id in (SELECT DISTINCT seg_id FROM %s WHERE (operation = 'DEL' AND user_id = $1 AND (current_timestamp > at_timestamp))) ", userSegTable, operationsTable)
-	row := r.db.QueryRow(query, usr_id)
-	if err := row.Scan(&id); id != 0 && err != nil {
-		return []string{}, err
+	// get list of unique seg_ids
+	seg_ids := []int{}
+	query := fmt.Sprintf("SELECT DISTINCT seg_id FROM %s WHERE user_id = $1", operationsTable)
+	_ = r.db.Select(&seg_ids, query, usr_id)
+
+	// check the last action performed
+	for _, segid := range seg_ids {
+		var temp string
+		query = fmt.Sprintf("SELECT operation FROM %s WHERE (user_id = $1 AND seg_id = $2) ORDER BY operation_id DESC LIMIT 1", operationsTable)
+		row := r.db.QueryRow(query, usr_id, segid)
+		if err := row.Scan(&temp); temp != "" && err != nil {
+			return []string{}, fmt.Errorf("Something went wrong, %s", err)
+		}
+
+		// if "DEL" -> remove user from the segment
+		if temp == "DEL" {
+			query := fmt.Sprintf("DELETE FROM %s WHERE user_id = $1 AND seg_id in (SELECT DISTINCT seg_id FROM %s WHERE (operation = 'DEL' AND user_id = $1 AND (current_timestamp > at_timestamp))) ", userSegTable, operationsTable)
+			row := r.db.QueryRow(query, usr_id)
+			if err := row.Scan(&id); id != 0 && err != nil {
+				return []string{}, err
+			}
+		}
 	}
 
 	// get user segments
-	query = fmt.Sprintf("SELECT DISTINCT seg_name FROM %s WHERE seg_id in (SELECT DISTINCT seg_id FROM %s WHERE (operation = 'ADD' AND user_id = $1 AND ((current_timestamp BETWEEN at_timestamp AND (at_timestamp + TTL * interval '1 second') OR TTL = 0))))", segmentsTable, operationsTable)
+	query = fmt.Sprintf("SELECT DISTINCT seg_name FROM %s WHERE seg_id in (SELECT DISTINCT seg_id FROM %s WHERE (user_id = $1))", segmentsTable, userSegTable)
 	err := r.db.Select(&slice, query, usr_id)
 	return slice, err
 }
@@ -36,12 +54,12 @@ func (r *UserPostgres) GetUserSegments(usr_id int) ([]string, error) {
 func (r *UserPostgres) AddUserToSegment(usr avito.User) (int, error) {
 	var s_id, id int
 
-	for _, x := range usr.Seg_names {
+	for _, segment := range usr.Seg_names {
 		// check if segment exists
 		query := fmt.Sprintf("SELECT seg_id FROM %s WHERE seg_name = ($1) LIMIT 1", segmentsTable)
-		row := r.db.QueryRow(query, x)
+		row := r.db.QueryRow(query, segment)
 		if err := row.Scan(&s_id); err != nil {
-			return 0, fmt.Errorf("Segment '%s' does not exist", x)
+			return 0, fmt.Errorf("Segment '%s' does not exist", segment)
 		}
 
 		// check if user was in a group
@@ -51,7 +69,7 @@ func (r *UserPostgres) AddUserToSegment(usr avito.User) (int, error) {
 		if err := row.Scan(&id); err != nil {
 			return 0, err
 		} else if int(id) > 0 {
-			return 0, fmt.Errorf("User '%d' is already in segment '%s'", usr.Id, x)
+			return 0, fmt.Errorf("User '%d' is already in segment '%s'", usr.Id, segment)
 		}
 
 		// create a new row
@@ -81,16 +99,16 @@ func (r *UserPostgres) AddUserToSegment(usr avito.User) (int, error) {
 	return 200, nil
 }
 
-func (r *UserPostgres) DeleteUserFromSegment(usr avito.User) (int, error) {
+func (r *UserPostgres) RemoveUserFromSegment(usr avito.User) (int, error) {
 	var s_id, id int
 
-	for _, x := range usr.Seg_names {
+	for _, segment := range usr.Seg_names {
 		// check if segment exists
 		query := fmt.Sprintf("SELECT seg_id FROM %s WHERE seg_name = ($1)", segmentsTable)
-		row := r.db.QueryRow(query, x)
+		row := r.db.QueryRow(query, segment)
 
 		if err := row.Scan(&s_id); err != nil {
-			return 0, fmt.Errorf("Segment '%s' does not exist", x)
+			return 0, fmt.Errorf("Segment '%s' does not exist", segment)
 		}
 
 		// check if user was in a group
@@ -100,7 +118,7 @@ func (r *UserPostgres) DeleteUserFromSegment(usr avito.User) (int, error) {
 		if err := row.Scan(&id); err != nil {
 			return 0, err
 		} else if int(id) == 0 {
-			return 0, fmt.Errorf("User '%d' is not in segment '%s'", usr.Id, x)
+			return 0, fmt.Errorf("User '%d' is not in segment '%s'", usr.Id, segment)
 		}
 
 		// delete all corresponding rows
@@ -138,8 +156,8 @@ func (r *UserPostgres) GetUserActions(usr avito.User) ([]string, error) {
 	query := fmt.Sprintf("SELECT \"user_id\", \"seg_id\", \"operation\", \"at_timestamp\" FROM %s WHERE (user_id = $1 AND EXTRACT(month FROM at_timestamp ) >= EXTRACT(month FROM $2 * interval '1 month') AND EXTRACT(year FROM at_timestamp) >= EXTRACT(year FROM $3 * interval '1 year'))", operationsTable)
 	err := r.db.Select(&temp, query, usr.Id, usr.Month, usr.Year)
 
-	for _, x := range temp {
-		s := strconv.Itoa(int(x.User_id)) + ";" + strconv.Itoa(int(x.Seg_id)) + ";" + x.Operation + ";" + x.At_timestamp + ";"
+	for _, action := range temp {
+		s := strconv.Itoa(int(action.User_id)) + ";" + strconv.Itoa(int(action.Seg_id)) + ";" + action.Operation + ";" + action.At_timestamp + ";"
 		slice = append(slice, s)
 	}
